New Development: No             JIRA Issue: Yes CAS-4094         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: sdmath unit test   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Bug fix on sdmath. The specunit and fluxunit (and telescopeparm) are 
properly handled in the task. Since specunit, frame, doppler don't affect 
an operation of the task, those parameters are just set to output data.
If fluxunit is set, scantable data can be overwritten depending on 
telescopeparm value. Thus, data is copied as needed to keep input data 
unchanged.

I added several unit tests to verify this fix.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/gcwrap@20092 ac9c0da5-131f-0410-9009-a9c195f183e6