New Development: No             JIRA Issue: Yes CAS-4429         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Describe your changes here...
sdimaging: added new parameters, convsupport, truncate, gwidth, and jwidth. 
sdgrid: parameter width is renamed as truncate, and added new parameters 
        convsupport, gwidth, and jwidth. 
convsupport is an integer parameter that specifies convolution support 
for SF gridding. truncate is a truncation radius for GAUSS and GJINC. 
gwidth is an HWHM of gaussian (effective for GAUSS and GJINC) while 
jwidth is a parameter for jinc function J_1(pi*x/c)/(pi*x/c), which is 
effective only for GJINC.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active/gcwrap@21782 ac9c0da5-131f-0410-9009-a9c195f183e6