New Development: No             JIRA Issue: Yes CAS-5496         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: parameter timerange is added          Test Programs: sdflag2 unit test   Put in Release Notes: No              Module(s): sd            Description: Describe your changes here...
New parameter timerange is added to sdflag2. It is implemented so that 
it works exactly same as in flagdata task. It supports all the formats 
described in cookbook, i.e., 'T0~T1', 'T0', 'T0+dT', '>T0', and '<T0', 
where T0 and T1 can be incomplete datetime string.

The parameter is effective when mode is either of 'manual', 'clip', or 
'interactive'. Unit test is added to check behavior of timerange.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/gcwrap@26380 ac9c0da5-131f-0410-9009-a9c195f183e6