New Development: No                 JIRA Issue:             Ready for Test: Yes          Interface Changes: No     What Interface Changed:              Test Programs:       Put in Release Notes: No                  Module(s):                Description: use fortran rules for complex mul/div Handling complex multiplication and division with correct treating of complex infinities (one element Inf regardless of the other) according to the C is complicated. E.g. a = NaN + 1e30 i; a * a is not NaN but a complex infinity (-Inf - NaN).
Treating this situation correctly has large performance impact. In GCC's
implementation it is about 4 times slower than the naive implementation, with
vectorization enabled the impact is even larger.
As correct treatment of complex infinfies when NaN appear in results is seldom
accounted for, or not required and most other languages do not have these
rules, the correct treatmeant can be disabled with the -fcx-fortran-rules flag.
This changes the semantics to those of the FORTRAN language which is removes
the need for rescuing the result when NaN appear. Python also follows FORTRAN
rules.
Additionally the correct behavior is not implemented in all compilers,
most notably clang which is the default compiler on MacOS. So turning off
correct treatment with GCC does not only make our code faster but also behave
the same on more compilers.

This has measurable impact on e.g. applycal where the cpu performance improves
by about 20%.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/gcwrap@32302 ac9c0da5-131f-0410-9009-a9c195f183e6