New Development: No             JIRA Issue: cas-7794         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
An earlier casacore mode (14 Aug 2015) added a parameter to the MSIter constructor
that has the MSIter keep the sort tables it creates in memory rather than storing
them on disk.  The default for the new parameter keeps the behavior using the older,
store on disk, semantics.  This mod replaces the ctor calls with explicit use of
the new parameter using the non-default value; the affected code will start using
in-memory sort tables.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/gcwrap@34292 ac9c0da5-131f-0410-9009-a9c195f183e6