Try to actually install in a reasonable way.
The current install code is written in a paradigm in which the destination is
a big CASA-private tree of everything. Let's not do that. Put Python stuff in
a vaguely correct directory and don't install the tests. Put XML files in a
subdirectory of share/. Try to namespace things sanely given all this.