New Development: Yes/No             JIRA Issue: CAS-9232         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_importasdm   Put in Release Notes: Yes              Module(s): importasdm, asdm2MS, mstool.asdmref            Description:
The "lazy" option in asdm filler (asdm2MS) used by importasdm has been extended
to work with FLOAT_DATA (ocorr_mode="ao" in importasdm).  

The asdmref function in mstool has also been extended to work with FLOAT_DATA
when present.  The interface has not changed and the user does not need to
know which column (DATA or FLOAT_DATA) is present in the MS.

During work on this issue it was discovered that the input data selection
options in asdm2MS (icm, isrt and its) are incompatible with the lazy option.
Any use of those options now will turn off any use of the lazy option at
the same time.  A warning is issued when that is done.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/gcwrap@39328 ac9c0da5-131f-0410-9009-a9c195f183e6